Add: Animaciones page, update: headers page.
Create animaciones page with a stateless widget that contains a Scaffold an in its child a stateful widget called Cuadrado Animado that return another stetefull widget called Cuadrado that returns a container that have a square shape, this is the preliminary to aminate the square.
Update headers page implementing animaciones page into the body of the scaffold.