Update: radial progress.
Update radial progress implemeting in the state the singletickerprovider to make the animation, add two new properties, one the animation controller and the other one a double called porcentajeAnterior that will save the initial percent, include the init state initializing the controller with a duration of 200 milliseconds and the dispose.
In the build method implementing the controller forward from 0.0 and create a new property called diferenciaAnimar that is the difference between widget.porcentaje and porcentajeAnterior also implementing a AnimatedBuilder that return the container with the custom paint.