Update: Radial progress, update: graficas circulares page.
Update radial progress adding 4 new properties, colorPrimario that aims to the color of the circle, colorSecundario that aims to the color of the arc that show the progress, grosorPrimario that aims to the thickness of the circle and grosorSecundario that aims to the thickness of the arc, implementing all the new properties into the painter.
Update graficas circulares page adding the arguments maked into the radial progress, the only one that is required is the porcentaje, every one else are optional, if the user dont give an argument the app will take a default value defined into the radial progress, also make the color of the button customizable but having a default value of colors.blue.