Update: graficas circulares page, update: radial progress.
Update graficas circulares page extracting the container and named it CustomRadialProgress, to this widget adding a new property called color that will used into the colorSecundario to change the color of the arc, in the body of the scaffold use a Column to show 4 radial progress into the screen with 4 diferents colors.
Update radial progress using the strokeCap into the arc paint to round the edges of the progress giving to it a better looking.