Update: Slideshow page
Update slideshow page turn the _Slides to a statefull widget to manage the state of the method, initialize a new instance of the PageController, also the initstate with the listener to know the index of the image where the user is,  also the dispose to save memory.
In the _Dot method create a new property called index to save the number of the index and by this way know which of the dot change color to identify the position of the image.
Implement the pageViewController into the PageView widget.