Add: Slider model, update: Pubspec.yaml.
Add Slider model to manage this class in a global level, in this class we will manage the index of the image with the property called currentPage, also making getters and setters, this is with the purpose to implement this class with provider.
Update pubspec.yaml installing provider from the pub.dev.