Update: Slideshow page
Update slideshow page wrapping the scaffold with the ChangeNotifierProvider because it contains the slides and the dots.
In the init state call the provider instance to update the current page when the user change the image.
In the dot class create a new property called pageViewIndex that calls the getter of the SliderModel with the provider, this property will be use to make a conditional with the index property, if the pageViewIndex is equal to the index, the color of the dot will be blue if not will be grey.