Update: Slideshow widget, update: slideshow page
Update slideshow widget adding a new property in the Slideshow class called slides that is a List of widgets and its contructor.
into the build method the _Slides class receive the list of slides as an argument.
In the _Slides class call the new property slides that is a list of widgets.
In the children of the PageView implement the widget.slides into a map and then to a list.
In the child of the container will receive the list of widgets instead of a svgpicture.

Update slideshow page implemeting five svgpictures from assets as a list of widgets.