Update: slideshow widget
Update slideshow widget adding a new property into the _Dots class called totalSlides that is a integer, it will save the number of widgets and will make dots as number of widgets.
In the children of the row implement a list generate that will take as lenght the totalSlides and as a generator the dot class in the index position.
And finally the _Dots class receive as an argument the totalSlides: slides.lenght to know how many widget are in the list of widget entered by the user.