Update: slideshow widget, update: slideshow page
Update slideshow widget adding four new properties, one is called puntosArriba to put the dots in the top or in the bottom of the screen, another one is bulletPrimario and bulletSecundario, bulletPrimario is to set the background color of the dot and bulletSecundario is to set the dot's color, the last one property is to set the dot's size.
The default values for puntosArriba is false, this will set the dows in the bottom of the screen, bulletPrimario defaults value is grey, and bulletSecundario is blue and the dotSize defaults value is 10.0.
All these new properties makes the widget totally customizable.
Update slideshow page giving values to all new properties created in the slideshow widget.