Update: slideshow widget.
Update slideshow widget renaming bulletPrimario and bulletSecundario to colorPrimario and colorSecundario also deleting the property dotSize because it will manage by the two new properties called bulletPrimario and bulletSecundario, the bulletPrimario will be a little bigger than the bulletSecundario it means that when the dot is selected the dot will be a little bigger than the other one dots.
Creating the properties into the _SlideshowModel to avoid these properties to use it with the provider.
The bulletPrimario will have a default value of 15 pixels and the bulletSecundario 12 pixels.