Add: pinterest menu, update: pinterest page
Add pinterest menu into the widgets folder, it will contain the class pinterest button with two properties, one is the onPressed that will contain the callback when the user press the icon and the second one is the icon.
The class pinterest menu will contain a list of items of PinterrestButton, in the beginning it will contain 4 pinterestbutton with its required onpressed and icons.
Update pinterest page using the pinterestmenu as the body of the scaffold.