Update: pinterest menu
Update pinterest menu  implementing the model to use provider, into the class of the model use a property to initialize the itemSeleccionado with its get and set
In the higher widget wrap the child to use the ChangeNotifierProvider and in its create the MenuModel.
Into the PinterestMenuButton create an instance of provider with the setter and getter to obtain the value.