Update: pinterest page, update: pinterest menu
Update pinterest page applying a stack widget into the body of the scaffold to put the grid and the menu over the grid, the PinterestMenu was wrapped with a positioned widget to apply a size in the bottom.
Update pinterest menu removing the center widget because it was apply only for develop purpose.