Update: pinterest page.
Update pinterest page turning the PinterestGrid to a StatefulWidget to avoid the program to use listeners, add two new properties, one is the controller that is an instance of ScrollController and the another one is called scrollAnterior to save the value of the controller.offset and make the conditional.
In the initstate add the listener and make the conditional to know if the menu has to be show or hide, before the conditional is ended, scrollAnterior takes the value of the controller.offset.
In the dispose implement controller.dispose to destroy the listener.