Update: pinterest page, update: pinterest menu
Update pinterest page implementing a model to use provider with a property boolean type called mostrar, this property will manage our menu if has to be show or hide depends on the scroll of the screen.
In the PinterestPage class extract the positiones widget calling it _PInterestMenuLocation, this with the purpose of have the provider in the highest widget.
Update pinterest menu with the property mostrar and its constructor, also in the PinterestMenu class wrap the child _PinterestMenuBackground with an AnimatedOpacity to show or hide the menu.