Update: pinterest menu, update: pinterest page.
Update pinterest menu adding three new properties to the class, one is the backgroundColor that is the backgroundColor of the menu, the second one is activeColor that is the color of the selected icon and the inactiveColor is the color of the no selected icons.
Implement the builder widget with the three new properties extracted from the provider.
Update pinterest page giving arguments to the properties created in the pinterest menu.