Update: headers widget, update: pubspec.
Update headers widget extracting the container of the IconHeader and called it _IconHeaderBackground, in the IconHeader implement a stack to put the _IconHeaderBackground, a positioned widget to put an icon in the background, a column to put 2 texts and an icon and another positioned to put an icon button in the top right of the screen.
Update pubspec installing font awesome to use another icons.