Update: boton gordo.
Update boton gordo implementing in the container of the background a cliprrect with a border radius to keep the icon inside the container, with Stack as a child and a positioned as a children.
Also implement the boton gordo information that is the information inside the background container, it returns a row with an icon, a text and an icon button.