Update: emergency page.
Update emergency page extracting the Boton Gordo widget in a new widget called BotonGordoTemp, also in the scaffold implement a stack to put all the boton gordo widget and the iconheader in the same page.
Implement a class called ItemBoton with the properties and the costructor implementing those properties, inside the build method put a list of the class ItemBoton, with the properties needed in the class ItemBoton, to turn the list in an instance of BotonGordo, use the map ah convert it to list.
Implement the map with the spread operator into the listview to create the buttons automatically.