Add: Routes, update: Launcher page.
Add routes into a new folder call routes, this new file is created to get the icon, title and page of our custom widgets designed in the last updates.
Update launcher page instancing the routes into the launcher page and using the pageRoutes to get the icon, title and page ahd show it in the listview of the page.