Add: Theme, update: main, update: launcher page.
Add theme into theme folder, this file contain the model to use provider, in this case will contain two properties that will managed the darkMode and the customTheme, implementing the getters and setters, in the setters make a logical change to just have one switch active.
Update main, wrapping the highgest widget of the app because the properties that will managed are the theme and the colors of the entire application.
Update launcher page, instancing the provider into the widget that contains the switches and giving the providers value to the switch.