Update: theme, update: main
Update theme in the class Theme Changer adding the property currentTheme into the model to use provider, create its getter and make a conditional into the setter of the darkMode and customTheme, also adding a parameter into a constructor of ThemeChanger using a switch conditional, this is with the purpose of make a user preferences.
Update main calling an instance of the provider to read the value and using it as argument into the theme.