Update: Launcher page, update: slideshow page.
Update launcher page using an instance of the provider and a property called accentColor to use the accentColor of the material, using it to change all the colors of the app depends on the user choise.
Update slideshow page using the same instance of provider and the same property accentColor to change the colors of the button, this with the purpose of give a better looking to the app.