Update: Graficas circulares page, update: headers page, update: pinterest page, update: headers widget.
Update graficas circulares page adding a multiplication into the customradialprogress to show the progress in different speeds, also adding an instance of provider to use the text color as the background color of the radial progress, when the theme is light the background color is black and when the theme is dark the background color is white.
Update headers page adding an instance of provider to call the colorScheme secondary, this property is called accentColor and is used as the required color of the custom widget HeaderWaves.
Update pinterest page adding the instance of provider to change the color of the application depends on the theme used.
Update headers widget adding the color property.