Update: theme, update: launcher page, update: main
Update theme erasing all the appbarTheme properties because it use the same color as the colorscheme secondary.
Update launcher page implementing the appBar backgroundColor from the colorScheme secondary.
Update main using the OrientationBuilder and returning the LauncherPage, the purpose of use OrientationBuilder is to know the position of the device, if is in portrait or landscape and by this way enhance the looking of the application in mobiles and tablets.