Update: Launcher page, add: Launcher Tablet Page, update: main
Update launcher page, adding to the title the word Mobile, this page will be show when the user is in a mobile device.
Add launcher tablet page, this page is a copy of launcher page with the difference that in the title is the word Tablet because this page will be show if the user is in a tablet or in a mobile device with a bigger screen size.
Update main adding the property screenSize that is an instance of MediaQuery to know the size of the screen, depends on the width size of the screen the app will show LauncherTabletPage or LauncherPage.