Add: layout model, update: launcher tablet page, update: main.
Add layout model to manage provider with the currentPage property.
Update launcher tablet page using the model of the layout to show the page when the user press the item calling the layoutmode currentpage adding the value of the pageRoutes[index].page.
Update main implementing MultiProvider because the mobile page will use ThemeChanger model and the tablet page will use the Layout Model.