Update: graficas circulares page, update: sliver list page, update: cuadrado animado page.
Update graficas circulares page changing the color of the floating action button from the color get into the constructor to the color of the theme, more exactly from the colorScheme secondary.
Update sliver list page adding a new bool property called isLarge, if the width of the screen is more than 500 pixels this property will be true, this property will be used to make a conditional in the minWidth of the matterial button, if isLarge is true it will apply size.width - 350, taking into account that the left menu in tablets has a width of 300 pixels and if isLarge is false it will apply size-width * 0.9.